requirements.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
python-decouple==3.8
pdfplumber==0.11.4
requests==2.32.3
beautifulsoup4==4.12.3
anthropic==0.40.0
openai==1.57.2
Pillow==10.4.0

# Transitive dependencies pinned to avoid installation issues
urllib3==2.6.3
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.11
soupsieve==2.8.3
anyio==4.12.1
sniffio==1.3.1
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
pydantic==2.12.5
pydantic_core==2.41.5
annotated-types==0.7.0
typing_extensions==4.15.0
typing-inspection==0.4.2
jiter==0.13.0
distro==1.9.0
tqdm==4.67.3
pdfminer.six==20231228
pypdfium2==5.5.0
cffi==2.0.0
pycparser==3.0
cryptography==46.0.5
PyJWT==2.11.0
sqlparse==0.5.5
asgiref==3.11.1
firecrawl-py==4.16.0
